% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALICE_pipeline.R
\name{ALICE_pipeline}
\alias{ALICE_pipeline}
\title{ALICE_pipeline}
\usage{
ALICE_pipeline(
  TCRgrObject,
  Q_val = 6.27,
  cores = 1,
  thres_counts = 1,
  N_neighbors_thres = 0,
  p_adjust_method = "BH",
  chain = "mouseTRB",
  stats = "OLGA",
  model = "-"
)
}
\arguments{
\item{TCRgrObject}{TCRgrapher object that contains clonoset table}

\item{Q_val}{selection factor. 1/Q sequences pass selection in a thymus. The
default value for mouses 6.27. If a human model is taken and Q is not changed
manually Q = 27 is used}

\item{cores}{number of used cores, 1 by default}

\item{thres_counts}{Only sequences with number of counts above this threshold
are taken into account}

\item{N_neighbors_thres}{Only sequences with number of neighbors above
threshold are used to calculate generation probability}

\item{p_adjust_method}{One of the methods from p.adjust from the stats package
possible options: "bonferroni", "holm", "hochberg", "hommel", "BH" or "fdr",
"BY", "none". "BH" is a default method.}

\item{chain}{Statistical model selection. Possible options: "mouseTRB",
"humanTRB", "humanTRA".}

\item{stats}{Tool that will be used for generation probability calculation.
Possible options: "OLGA", "SONIA". "SONIA" also calculates Q for every sequence.}

\item{model}{Standard OLGA generation probability model is used by default.
To set your one generation probability model, write "set_custom_model_VDJ
<path_to_folder_with_model>". A generation probability model is usually IGOR output.
A folder should contain the following files: V_gene_CDR3_anchors.csv,
J_gene_CDR3_anchors.csv, model_marginals.txt, model_params.txt. Some models
can be found in the folder "model"}
}
\value{
Function returns TCRgrapher object filtered by number
of counts and number of neighbors with additional columns. Additional columns
are the following
\itemize{
\item{"ALICE.D"}{"Number of neighbors in clonoset. Neighbor is a similar sequence
with one mismatch"}
\item{"ALICE.VJ_n_total"}{"Number of unique sequences with given VJ combination"}
\item{"ALICE.Pgen"}{"Probability to be generated computed by OLGA"}
\item{"ALICE.p_value"}{"p value under null hypothesis that number of sequence's
neighbors is not more than in the random model"}
\item{"ALICE.p_adjust"}{"p value with multiple testing correction"}
\item{"ALICE.log_p_value"}{"log of p_value. p values are usually very small
and can be rounded to zero"}
}
}
\description{
The function takes a TCRgrapher object as an input and performs neighborhood
enrichment analysis using the ALICE algorithm. See ?TCRgrapher.
}
