% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{C57BL6_MOUSE_TRB}
\alias{C57BL6_MOUSE_TRB}
\title{VJ combinations probabilities for mouse TRB model obtained from C57BL/6
unproductive sequences}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 35 rows and 14 columns.
}
\source{
\url{https://github.com/KseniaMIPT/tcrgrapher/tree/master/models/C57BL.6_mouseTRB}
}
\usage{
C57BL6_MOUSE_TRB
}
\description{
A data frame with probabilities of VJ combinations
All probabilities based on unproductive sequences from C57BL/6 mice line
}
\author{
Ksenia R. Lupyr
}
\keyword{datasets}
