% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist3.R
\name{calc_TCRdist3_radius}
\alias{calc_TCRdist3_radius}
\title{calc_TCRdist3_radius}
\usage{
calc_TCRdist3_radius(
  TCRgrObject,
  cores = 1,
  organism = "mouse",
  chain = "beta"
)
}
\arguments{
\item{TCRgrObject}{See ?TCRgrapher}

\item{cores}{the number of cores to use. The default value is 1.}

\item{organism}{Possible options: mouse, human. The default value is "mouse"}

\item{chain}{Possible options: alpha, beta. The default value is "beta"}
}
\value{
TCRgrapher object. A clonoset contains additional column "tcrdist3.radius"
}
\description{
The function takes TCRgrapher object as an input and finds the optimal radius
with tcrdist3 python package.
}
