% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR_pipeline.R
\name{create_count_table}
\alias{create_count_table}
\title{create_count_table}
\usage{
create_count_table(dir, metadata, v_gene = TRUE)
}
\arguments{
\item{dir}{path to the directory where the clonotype tables are located}

\item{metadata}{table should contain 'file' column with names of the files
in directory specified earlier. Samples should contain 'cdr3aa' and 'count'
columns.}
}
\value{
Function returns count table with unique amino acid sequences as
row names and sample names as column names
}
\description{
Function takes clonotype tables from specified directory and creates a table
where each row corresponds to unique amino acid clonotype and each column
corresponds to the sample.
}
