% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR_pipeline.R
\name{filter_edgeR_res}
\alias{filter_edgeR_res}
\title{filter_edgeR_res}
\usage{
filter_edgeR_res(res_dt)
}
\arguments{
\item{res_dt}{the output of edgeR_pipeline function}
}
\value{
subset of res_dt with only 'vs all' comparisons and additional columns
}
\description{
The function takes "vs all" comparisons and checks if all pairwise comparisons
are consistent with the given "vs all" comparison. 'the_worst_pairwise_p' column
shows the worst p value in all pairwise comparisons.
}
