% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_capabilities.R
\name{find_TCR_components}
\alias{find_TCR_components}
\title{find_TCR_components}
\usage{
find_TCR_components(clonoset, g)
}
\arguments{
\item{clonoset}{clonoset table. To get from TCRgrapher object use clonoset(\if{html}{\out{<TCRgrapher object>}})}

\item{g}{make_TCR_graph output}
}
\value{
Function returns the same clonoset table with additional column "cluster_id"
}
\description{
Function takes clonoset table and returns the same table with additional
column "cluster_id". All clusters of neighbors with one mismatch have unique
id. Function uses "components" function from igraph package.
}
