% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igor.R
\name{make_gen_model}
\alias{make_gen_model}
\title{make_gen_model}
\usage{
make_gen_model(
  nonfunc_path = "test",
  command = "igor",
  wd = "igor_output",
  batch_name = "new_inference",
  species = "mouse",
  chain = "beta"
)
}
\arguments{
\item{nonfunc_path}{A path to the file with nonfunctional nucleotide sequences.
The file should contain only sequences (one sequence per line).}

\item{command}{a command or a path to run on the command line. The default
value is 'igor'. For example, the user can specify '/user/.local/bin/igor'.}

\item{wd}{a name of the working directory. The default value is 'igor_output'.}

\item{batch_name}{The default value is 'new_inference'}

\item{species}{Possible values: mouse, human. If "human" is set, the default
IGoR genomic template is used. If "mouse" is set, the default TCRgrapher
genomic template is used. See Details.}

\item{chain}{Possible values for human model: alpha, beta. Possible values for
mouse model: beta.If "human" is set, the default IGoR genomic template is used.
If "mouse" is set, the default TCRgrapher genomic template is used. See Details.}
}
\description{
The function uses the IGoR software to infer a new generation model. IGoR should
be installed. The recommended version is v1.3.0. The function takes the path to
the file with non-functional nucleotide CDR3 sequences and writes a generation
model based on these sequences to the working directory. To use more features,
one can use IGoR directly. See the documentation at https://qmarcou.github.io/IGoR/#version.
}
\details{
Additional details...How tcgrapher genomic template was made...# TODO
}
