% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_capabilities.R
\name{make_TCR_graph}
\alias{make_TCR_graph}
\title{make_TCR_graph}
\usage{
make_TCR_graph(df)
}
\arguments{
\item{df}{output of tcrgrapher function}
}
\value{
Function returns an igraph graph object
}
\description{
Function makes graph from tcrgrapher output with igraph package. Every node
of the graph is an unique clonotype from the table (one line). Edges
connects clonotypes with one amino acid mismatch or identical clonotypes if
they were in separate lines.
}
