% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLGA_J}
\alias{OLGA_J}
\title{TRBJ usage probabilities}
\format{
An object of class \code{numeric} of length 14.
}
\source{
\url{https://github.com/statbiophys/OLGA/tree/master/olga/default_models/mouse_T_beta}
}
\usage{
OLGA_J
}
\description{
Vectors with probabilities of TRBJ usage
All probabilities based on OLGA mouse T beta model
}
\author{
Pavel V. Shelyakin
}
\keyword{datasets}
