% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLGA_V}
\alias{OLGA_V}
\title{TRBV usage probabilities}
\format{
An object of class \code{numeric} of length 35.
}
\source{
\url{https://github.com/statbiophys/OLGA/tree/master/olga/default_models/mouse_T_beta}
}
\usage{
OLGA_V
}
\description{
Vectors with probabilities of TRBV usage
All probabilities based on OLGA mouse T beta model
}
\author{
Pavel V. Shelyakin
}
\keyword{datasets}
