% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLGAVJ_HUMAN_TRA}
\alias{OLGAVJ_HUMAN_TRA}
\title{VJ combinations probabilities for human TRA model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 47 rows and 61 columns.
}
\source{
\url{https://github.com/pogorely/ALICE/blob/master/OLGA_A_hum_alpha.rda}
}
\usage{
OLGAVJ_HUMAN_TRA
}
\description{
A data frame with probabilities of VJ combinations
All probabilities based on OLGA human T beta model
}
\keyword{datasets}
