% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLGAVJ_HUMAN_TRB}
\alias{OLGAVJ_HUMAN_TRB}
\title{VJ combinations probabilities for human TRB model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 59 rows and 13 columns.
}
\source{
\url{https://github.com/pogorely/ALICE/blob/master/OLGA_V_J_hum_beta.rda}
}
\usage{
OLGAVJ_HUMAN_TRB
}
\description{
A data frame with probabilities of VJ combinations
All probabilities based on OLGA human T beta model
}
\keyword{datasets}
