% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OLGAVJ_MOUSE_TRB}
\alias{OLGAVJ_MOUSE_TRB}
\title{VJ combinations probabilities for mouse TRB model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 35 rows and 14 columns.
}
\source{
\url{https://github.com/statbiophys/OLGA/tree/master/olga/default_models/mouse_T_beta}
}
\usage{
OLGAVJ_MOUSE_TRB
}
\description{
A data frame with probabilities of VJ combinations
All probabilities based on OLGA mouse T beta model
}
\author{
Pavel V. Shelyakin
}
\keyword{datasets}
