% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GLIPH2.R
\name{run_GLIPH2}
\alias{run_GLIPH2}
\title{run_GLIPH2}
\usage{
run_GLIPH2(
  TCRgrObject,
  gliph2_path,
  control = "TCRgrapher",
  kmer_min_depth = 1e+08
)
}
\arguments{
\item{TCRgrObject}{TCRgrapher object}

\item{gliph2_path}{path to the GLIPH2 executable file}

\item{control}{The possible values are 'TCRgrapher' and 'GLIPH2'. The default
value is 'TCRgrapher'. If you choose 'GLIPH2' the standard GLIPH2 control from the
server will be used. If you would like to use the control generated in our lab,
choose 'TCRgrapher'}

\item{kmer_min_depth}{One of the GLIPH2 parameters. The default value is 100000000.
We recommend not using clusters based on k-mers. The default value is chosen
to avoid its computation. If you want to use them, choose the small value
(kmer_min_depth = 3, for example). Keep in mind that it will also increase
computational time.}
}
\value{
The function returns the same TCRgrapher object with additional columns
in the clonotype table: 'min_gliph2_fisher_score_global' and 'min_gliph2_fisher_score_(k)mer'
if k-mers were present.
}
\description{
The function takes the TCRgrapher object as an input and runs GLIPH2 analysis.
In GLIPH2 output, one sequence can be included in different clusters and
have different p-values. At the last step, the function assigns the minimum p-value
for every clonotype from the TCRgrapher clonotype table. If there are different types
of patterns in the analysis, the output will have a separate column for each
type of pattern. The GLIPH2 executable file must be installed.
See details at http://50.255.35.37:8080/tools
}
