% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRgrapherObject.R
\name{subset,TCRgrapher-method}
\alias{subset,TCRgrapher-method}
\title{#' @export
setGeneric("subset", function(x, samples) standardGeneric("subset"))
Subseting for TCRgrapher objects}
\usage{
\S4method{subset}{TCRgrapher}(x, samples)
}
\arguments{
\item{x}{TCRgrapher object}

\item{subset}{Vector with sample ids that should be kept}
}
\description{
The function takes subset from clonotype table, metadata and edges if they are
present. Samples to keep should be specified. 'clone_id' column will be updated.
}
