% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRgrapherCountsObject.R
\name{take_subset_from_count_table}
\alias{take_subset_from_count_table}
\title{take_subset_from_count_table}
\usage{
take_subset_from_count_table(count_table, samples)
}
\arguments{
\item{count_table}{a data frame where each row corresponds to unique
amino acid clonotype and each column corresponds to the sample.}

\item{samples}{a vector of names that matches the names of the columns to be
taken into analysis}
}
\value{
count_table with specified columns and without rows with zero counts
}
\description{
If you would like to take part of the samples into analysis, it is important
to take a subset from a count table correctly, because number of rows influence
the result.
}
\examples{
# metadata_CD8 <- metadata[metadata$Cell_Population == 'CD8',]
# count_table_CD8 <- take_subset_from_count_table(count_table, metadata_CD8$samples)
}
