% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRdist3.R
\name{TCRdist3_radius}
\alias{TCRdist3_radius}
\title{TCRdist3_radius}
\usage{
TCRdist3_radius(TCRgrObject, cores)
}
\arguments{
\item{TCRgrObject}{See ?TCRgrapher}

\item{cores}{the number of cores to use}
}
\description{
The function takes TCRgrapherObject as an input and finds the optimal radius
with tcrdist3 python package.
}
