% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VDJT}
\alias{VDJT}
\title{All combinations of V and J segments}
\format{
A data frame with 253 rows and 2 variables
\itemize{
\item V TRBV segment name
\item J TRBJ segment name
}
}
\usage{
VDJT
}
\description{
A dataset containing all combinations of V and V segments with V segments'
names in the first column and J segments' names in the second column.
}
\author{
Pavel V. Shelyakin
}
\keyword{datasets}
