% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR_pipeline.R
\name{wilcox_pipeline}
\alias{wilcox_pipeline}
\title{wilcox_pipeline}
\usage{
wilcox_pipeline(TCRgrObject, comparison)
}
\arguments{
\item{comparison}{a name of the column that specifies levels of comparison in the metadata}

\item{TCRgrCounts}{a TCRgrapherCounts object. For more details see ?TCRgrapherCounts}
}
\value{
data.table with statistics for every clonotype or cluster of clonotypes
performed for each pair of groups and for every group vs. all others.
'comparison' column shows comparisons in the format 'Group1 vs Group2'.
P-values correspond to the alternative hypothesis that the median of the first
group is higher than the median of the second group.
}
\description{
The function performs all pairwise comparisons between groups and compares
each group vs all others using wilcox test to identify significantly
expanded clonotypes of clusters of clonotypes.
}
