% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_search_bfs.R
\name{bfs_for_TCRs}
\alias{bfs_for_TCRs}
\title{bfs_for_TCRs}
\usage{
bfs_for_TCRs(
  clonoset,
  src,
  comp_id,
  v_gene = TRUE,
  j_gene = FALSE,
  edges = c()
)
}
\arguments{
\item{clonoset}{To get clonoset use clonoset(your_TCR_grapher_object)}

\item{src}{clone_id of the source clonotype}

\item{comp_id}{cluster_id for the visited vertexes. All other vertexes will have
-1 cluster_id if there are none. If some cluster_ids have already been set,
they will be kept.}

\item{edges}{matrix to which new edges will be added}
}
\value{
The function returns a list with a clonoset where all neighbors of the source
clonotype have the given id. The second item of the list is the matrix with edges.
}
\description{
The function searches for all neighbors of the source (src) clonotype. A neighbor
is a clonotype that differs by one amino acid mismatch.
}
