% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR_pipeline.R
\name{filter_wilcox_res}
\alias{filter_wilcox_res}
\title{filter_wilcox_res}
\usage{
filter_wilcox_res(res_dt, comparison)
}
\arguments{
\item{res_dt}{the output of wilcox_pipeline function}

\item{comparison}{a name of the column that specifies levels of comparison}
}
\value{
subset of res_dt with only 'vs all' comparisons and additional column
'consistent'
}
\description{
The function takes "vs all" comparisons and checks if all pairwise comparisons
are consistent with the given "vs all" comparison.
}
