% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLGA_pipeline.R
\name{find_cluster}
\alias{find_cluster}
\title{find_cluster}
\usage{
find_cluster(df)
}
\arguments{
\item{df}{data.table}
}
\value{
the same data.table with additional column 'cluster_id'
}
\description{
Function gives unique number (cluster_id) to all similar sequences with one
mismatch it doesn't depend on VJ combination. Sequences with distance more
than one mismatch can have one cluster_id if they have common neighbor
between them
}
