% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_capabilities.R
\name{find_TCR_components}
\alias{find_TCR_components}
\title{find_TCR_components}
\usage{
find_TCR_components(df, g)
}
\arguments{
\item{df}{output of tcrgrapher function}

\item{g}{make_graph output}
}
\value{
Function returns the same data.table with additional column
"cluster_id"
}
\description{
Function takes tcrgrapher output and returns the same table with additional
column "cluster_id". All clusters of neighbours with one mismatch have unique
id. Function uses "components" function from igraph package.
}
