% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_search_bfs.R
\name{find_TCR_components_by_bfs}
\alias{find_TCR_components_by_bfs}
\title{find_TCR_components_by_bfs}
\usage{
find_TCR_components_by_bfs(clonoset, v_gene = TRUE, j_gene = FALSE)
}
\arguments{
\item{clonoset}{To get clonoset use clonoset(your_TCR_grapher_object)}

\item{v_gene}{Boolean value. If 'v_gene' is 'TRUE', only clonotypes with the
same V genes will be part of one connectivity component. Default value is 'TRUE'.}

\item{j_gene}{Boolean value. If 'j_gene' is 'TRUE', only clonotypes with the
same J genes will be part of one connectivity component. Default value is 'FALSE'.}
}
\value{
The function returns the clonoset with the new column "cluster_id", where
every connectivity component has a unique value.
}
\description{
The function searches for connectivity components of the graph, where every node is a
clonotype with a unique clone_id and edges connect clonotypes that differ by one
amino acid mismatch.
}
