% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR_pipeline.R
\name{heatmap_expanded}
\alias{heatmap_expanded}
\title{heatmap_expanded}
\usage{
heatmap_expanded(TCRgrCounts, expanded_test_res, comparison)
}
\arguments{
\item{TCRgrCounts}{TCRgrapherCounts object. For more details see ?TCRgrapherCounts}

\item{expanded_test_res}{table with features (clonotypes or clusters of clonotypes)}

\item{comparison}{a name of the column that specifies levels of comparison in the metadata
to draw. It could be an output of edgeR_pipeline or wilcox_pipeline.}
}
\description{
The function takes an output of edgeR_pipeline or wilcox_pipeline functions
and creates a heatmap using the ComplexHeatmap library. It is recommended
to use filter_edgeR_res or filter_wilcox_res previously and filter results by
some threshold.
}
