% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_capabilities.R
\name{make_TCR_graph}
\alias{make_TCR_graph}
\title{make_TCR_graph}
\usage{
make_TCR_graph(clonoset, v_gene = TRUE, j_gene = FALSE)
}
\arguments{
\item{clonoset}{clonoset clonoset table. To get from TCRgrapher object use clonoset(\if{html}{\out{<TCRgrapher object>}})}
}
\value{
Function returns an igraph graph object
}
\description{
Function makes a graph from the clonoset table with the igraph package. Every node
of the graph is an unique clonotype from the table (one line). Edges
connects clonotypes with one amino acid mismatch or identical clonotypes if
they are in separate lines.
}
