% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLGA_pipeline.R
\name{pipeline_OLGA}
\alias{pipeline_OLGA}
\title{Main function}
\usage{
pipeline_OLGA(
  df,
  Q = 6.27,
  cores = 1,
  prompt = F,
  Read_thres = 0,
  Read_thres2 = 1,
  N_neighbors_thres = 1
)
}
\arguments{
\item{df}{data.table}

\item{Q}{selection factor}

\item{cores}{number of cores}

\item{prompt}{smth}

\item{Read_thres}{threshold 1}

\item{Read_thres2}{threshold 2}

\item{N_neighbors_thres}{threshold 3}
}
\description{
Main function
}
