% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLGA_pipeline.R
\name{pipeline_OLGA}
\alias{pipeline_OLGA}
\title{Main function}
\usage{
pipeline_OLGA(
  df,
  Q = 6.27,
  cores = 1,
  prompt = F,
  Read_thres = 0,
  Read_thres2 = 1,
  N_neighbors_thres = 1
)
}
\arguments{
\item{df}{data.table}
}
\description{
Main function
}
