% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLGA_pipeline.R
\name{pipeline_OLGA}
\alias{pipeline_OLGA}
\title{pipeline_OLGA}
\usage{
pipeline_OLGA(
  df,
  Q = 6.27,
  cores = 1,
  thres_counts = 1,
  N_neighbors_thres = 1,
  p_adjust_method = "BH"
)
}
\arguments{
\item{df}{data.table}

\item{Q}{selection factor. 1/Q sequences pass selection in thymus. Default
value for mouses 6.27}

\item{cores}{number of used cores, 1 by default}

\item{thres_counts}{Only sequences with number of counts above this threshold
are taken into account}

\item{N_neighbors_thres}{Only sequences with number of neighbors above
threshold are used to calculate generation probability}

\item{p_adjust_method}{one of the method from p.adjust from stats package
possible options: "bonferroni", "holm", "hochberg", "hommel", "BH" or "fdr",
"BY", "none". "BH" is a default method.}
}
\value{
Function returns the same table that was in input filtered by number
of counts and number of neighbors with additional columns. Additional columns
are the following
\itemize{
\item{"D"}{"Number of neighbors in clonoset. Neighbor is a similar sequence
with one mismatch"}
\item{"VJ_n_total"}{"Number of unique sequences with given VJ combination"}
\item{"Pgen"}{"Probability to be generated computed by OLGA"}
\item{"Pgen_sum_corr"}{"Sum of Pgen of all sequences similar to the given
with one mismatch"}
\item{"Pgen_by_VJ"}{"Conditional probability. Sum of probabilities to be
generated with given VJ combination"}
\item{"p_val"}{"p value under null hypothesis that number of sequence's
neighbors is not more than in the random model"}
\item{"p_adjust"}{"p value with multiple testing correction"}
}
}
\description{
Main fucntion that takes table with cdr3 sequences as an input. Table should
have the following columns (names of the colums are not important but the
followig order is necessary)
\itemize{
\item{"Read.count"}{"Number of unique reads per cdr3 sequence"}
\item{"freq"}{"Clonotype frequency in the clonoset"}
\item{"cdr3nt"}{"CDR3 nucleotide sequence"}
\item{"bestVGene"}{"TRBV segment"}
\item{"bestVGene"}{"TRBD segment"}
\item{"bestJGene"}{"TRBJ segment"}
\item{"VEnd"}{"Position of the end of V segment in CDR3 sequence"}
\item{"DStart""}{"Position of the start of D segment in CDR3 sequence"}
\item{"DEnd"}{"Position of the end of D segment in CDR3 sequence"}
\item{"JStart"}{"Position of the start of J segment in CDR3 sequence"}
}
}
