% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALICE_pipeline.R
\name{pval_with_abundance}
\alias{pval_with_abundance}
\title{pval_with_abundance}
\usage{
pval_with_abundance(df)
}
\arguments{
\item{df}{output of tcrgrapher function}
}
\value{
Function returns the same data.table with additional columns:
\itemize{
\item{"pval_with_abundance_log2_counts"}{"Recalculated p-value considering
count number of every clonotype. Log2 is used for count normalization"}
\item{"pval_with_abundance_counts"}{"Recalculated p-value considering
count number of every clonotype. There is no count normalization"}
}
\description{
Function calculates p-value taking into account abundance of every clonotype
}
