% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TCRNET.R
\name{run_TCRNET}
\alias{run_TCRNET}
\title{run_TCRNET}
\usage{
run_TCRNET(
  TCRgrObject,
  background_path,
  command = "vdjtools",
  search_scope = c(1, 0, 1),
  grouping = "vj"
)
}
\arguments{
\item{background_path}{A path to the background sample}

\item{command}{A command or a path to run on the command line. The default
value is 'vdjtools'. For example, the user can specify 'java -jar /your_path/vdjtools-1.2.1.jar'}

\item{search_scope}{A vector with three numbers: allowed number of substitutions (s),
indels (id) and total number of mismatches (t). The default value is c(1,0,1)}

\item{grouping}{The default value is 'vj'. Possible values: 'vj', 'vjl', 'v', 'dummy'}
}
\description{
The function runs TCRNET (CalcDegreeStats command from vdjtools). VDJtools must
be installed. See documentation at https://vdjtools-doc.readthedocs.io/en/master/
}
