% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRgrapherObject.R
\name{subset,TCRgrapher-method}
\alias{subset,TCRgrapher-method}
\title{Subseting for TCRgrapher objects}
\usage{
\S4method{subset}{TCRgrapher}(x, subset)
}
\arguments{
\item{x}{TCRgrapher object}

\item{subset}{Vector with sample ids that should be kept}
}
\description{
The function takes subset from both clonotype table and metadata. Samples that
to keep should be specified.
}
