% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRgrapherObject.R
\name{TCRgrapher}
\alias{TCRgrapher}
\title{TCRgrapher object}
\usage{
TCRgrapher(
  files_path,
  count_column,
  cdr3nt_column,
  cdr3aa_column,
  v_gene_column,
  j_gene_column,
  metadata_path = NA,
  files_column = 1,
  ids_column = 2
)
}
\arguments{
\item{files_path}{Path to the file with a clonoset or path to the directory
with files. The file should contain a header and the following values: count,
cdr3nt, cdr3aa, V gene and J gene. The position of each column should be specified further.
Indexing is 1-based. Additional columns will be kept.}

\item{count_column}{Integer value that specifies the position of the count column in the clonoset.}

\item{cdr3nt_column}{Integer value that specifies the position of the cdr3nt column in the clonoset.}

\item{cdr3aa_column}{Integer value that specifies the position of the cdr3aa column in the clonoset.}

\item{v_gene_column}{Integer value that specifies the position of the V gene column in the clonoset.}

\item{j_gene_column}{Integer value that specifies the position of the J gene column in the clonoset.}

\item{metadata_path}{Path to the metadata file. The file must contain a header,
column with file names in the directory specified earlier, and a column with an unique
name for every file. The default value is NA. If the metadata path is not specified,
metadata will be generated.}

\item{files_column}{Integer value that specifies the position of the column with
file names in metadata}

\item{ids_column}{Integer value that specifies the position of the column with
unique names for every file in metadata}
}
\description{
TCRgrapher is a S4 class that can be constructed by calling the TCRgrapher( )
function. See "Details". TCRgrapher contains clonoset and metadata as
data.tables. You can get them by calling metadata(x) or clonoset(x).
}
\details{
There are three ways to initialize an object.

(1) Specify the path to the one file without metadata

If the path for only one file is specified, metadata will be produced automatically
and will contain one row and two columns: "file" and "sample_id".
Clonoset will have an additional column "sample_id" with one unique value and a
column "clone_id" with an unique id for every row.
Positions of "count", "cdr3nt", "cdr3aa", "V gene", "J gene" clonoset's columns must be specified.

(2) Specify the path to the directory with files without metadata

If the path for the directory with files is specified, and metadata is not specified
metadata will be produced automatically and will have a number of rows equal
to the number of samples in the directory. All files will be merged into one
data.table with additional columns "sample_id" and "clone_id" with
unique ids for every row. Positions of "count", "cdr3nt", "cdr3aa",
"V gene", "J gene" clonoset's columns must be specified.

(3) Specify the path to the directory with files and metadata

If the path for the directory with files and the path to the metadata are specified,
only files from the metadata column "file" will be taken into consideration.
Clonosets will be merged into one data.table with additional columns "sample_id"
and "clone_id" with unique ids for every row.
Values in the "sample_id" column will be the same as in the "sample_id"
metadata column. Positions of "count", "cdr3nt", "cdr3aa", "V gene", "J gene" clonoset's
columns and "file", "sampl_id" metadata's columns must be specified.

TCRgrapherCounts is a subclass of TCRgrapher that store additional data:
count_table and feature_info table
}
\examples{
# # (1) Path to the one file without metadata
# file_path <- paste0(find.package('tcrgrapher'), "tests/testthat/testdata/clonosets_vdjtools_format.tsv")
# TCRgrObject <- TCRgrapher(file_path, 1, 3, 4, 5, 7) # positions of clonoset's columns
# # metadata has the same values in both columns. Let's change one of them
# metadata(TCRgrObject)[1,'sample_id'] <- 'sample_1'

# # (2) Path to the directory with files without metadata
# dir_path <- paste0(find.package('tcrgrapher'), "tests/testthat/testdata/test_dir_with_clonosets")
# TCRgrObject <- TCRgrapher(dir_path, 1, 3, 4, 5, 7)

# # (3) Path to the directory with files and metadata
# dir_path <- paste0(find.package('tcrgrapher'), "tests/testthat/testdata/test_dir_with_clonosets")
# metadata_path <- testthat::test_path("testdata/metadata.tsv")
# TCRgrObject <- TCRgrapher(dir_path, 1, 3, 4, 5, 7, # positions of clonoset's columns
                        #  metadata_path, 1, 2)      # positions of metadtata's columns


}
\seealso{
#TODO
}
