% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRgrapherCountsObject.R
\name{TCRgrapherCounts}
\alias{TCRgrapherCounts}
\title{TCRgrapherCounts object}
\usage{
TCRgrapherCounts(TCRgrObject, v_gene = TRUE, j_gene = FALSE)
}
\arguments{
\item{TCRgrObject}{TCRgrapher object. It can be constructed by calling the
TCRgrapher( ) function.}

\item{v_gene}{Boolean value. If 'v_gene' is 'TRUE', clonotypes with the same amino
acid sequences but different V genes are presented in different rows.
Default value is 'TRUE'.}

\item{j_gene}{Boolean value. If 'j_gene' is 'TRUE', clonotypes with the same amino
acid sequences but different J genes are presented in different rows.
Default value is 'FALSE'.}
}
\value{
Function returns TCRgrapherCounts object that contains clonoset and
metadata from TCRgrapher object, count table and feature info table
}
\description{
Function takes a TCRgrapher object as an input and creates a count table
where each row corresponds to an unique amino acid sequence and each column
corresponds to the sample.
}
